+++ tests/analysis/test_behavior.py
@@ -8,7 +8,6 @@ from typing import Dict, List, Optional, Union
 import numpy as np
 import pandas as pd
 import pytest
-
 from shapely.geometry import Point
 
 from zebtrack.analysis.behavior import BehavioralAnalyzer, ConcreteBehavioralAnalyzer